fix: move color picker to separate component
