+++ package.json
@@ -28,7 +28,6 @@
     "next-plausible": "^3.12.2",
     "react": "^18",
     "react-code-block": "^1.0.0",
-    "react-color": "^2.19.3",
     "react-color-palette": "^7.3.0",
     "react-dom": "^18",
     "react-hook-form": "^7.53.0",
