+++ src/app/components/InboxDesignForm.tsx
@@ -1,9 +1,8 @@
 "use client";
 
-import React, { useRef, useState, useEffect } from "react";
+import React from "react";
+import dynamic from "next/dynamic";
 import { useTheme } from "../contexts/ThemeContext";
-import { ColorPicker, useColor } from "react-color-palette";
-import "react-color-palette/css";
 import { Select } from "chakra-react-select";
 import {
   VStack,
@@ -13,12 +12,13 @@ import {
   Switch,
   SimpleGrid,
   Input,
-  Box,
   Flex,
 } from "@chakra-ui/react";
-import { PickerIcon } from "./icons/Picker";
-import useClickOutside from "../hooks/useClickOutside";
-import { UseFormRegister, UseFormSetValue, Controller } from "react-hook-form";
+import { Controller } from "react-hook-form";
+
+const ColorPickerField = dynamic(() => import("./ColorPickerField"), {
+  ssr: false,
+});
 
 const items = [
   {
@@ -67,70 +67,6 @@ const items = [
     placeholder: "black",
   },
 ];
-
-function ColorPickerField({
-  register,
-  setValue,
-  placeholder,
-  name,
-  label,
-  index,
-}: {
-  register: UseFormRegister<any>;
-  setValue: UseFormSetValue<any>;
-  placeholder: string;
-  name: string;
-  label: string;
-  index: number;
-}) {
-  const [color, setColor] = useColor(placeholder);
-  const [openColorPicker, setOpenColorPicker] = useState(false);
-  const pickerPosition = index % 2 === 0 ? { left: 0 } : { right: 0 };
-  const pickerRef = useRef<HTMLDivElement>(null);
-
-  useClickOutside([pickerRef], () => setOpenColorPicker(false));
-
-  useEffect(() => {
-    setValue(name, color.hex);
-  }, [color, name, setValue]);
-
-  const handleColorChange = (newColor: any) => {
-    setColor(newColor);
-    setValue(name, newColor.hex);
-  };
-
-  return (
-    <FormControl position="relative">
-      <FormLabel fontSize="sm" color="white" opacity={0.6}>
-        {label}
-      </FormLabel>
-      <Box position="relative" ref={pickerRef}>
-        <Input
-          {...register(name)}
-          placeholder={placeholder}
-          value={color.hex}
-          size="sm"
-          readOnly
-          onClick={() => setOpenColorPicker(!openColorPicker)}
-          cursor="pointer"
-        />
-
-        <PickerIcon className="inset-y-2 right-3 absolute pointer-events-none" />
-
-        {openColorPicker && (
-          <Box position="absolute" top="100%" paddingTop="10px" zIndex={50} {...pickerPosition}>
-            <ColorPicker
-              color={color}
-              onChange={handleColorChange}
-              hideInput={["rgb", "hsv"]}
-              height={100}
-            />
-          </Box>
-        )}
-      </Box>
-    </FormControl>
-  );
-}
 
 export function InboxDesignForm() {
   const {
